+++ example/lib/shared/services/theme_service_hive.dart
@@ -84,6 +84,7 @@ class ThemeServiceHive implements ThemeService {
     Hive.registerAdapter(TabBarIndicatorSizeAdapter());
     Hive.registerAdapter(AdaptiveThemeAdapter());
     Hive.registerAdapter(FlexPaletteTypeAdapter());
+    Hive.registerAdapter(VisualDensityEnumAdapter());
   }
 
   // ----------